updating

+++ config.py
@@ -7,10 +7,10 @@ if USE_JAX:
     config.update("jax_enable_x64", True)
     import jax.numpy as jnp
     K = jnp
-    dtype = jnp.float64
+    dtype = jnp.float32
     print('Using JAX backend')
 else:
     import numpy as np
     K = np
-    dtype = np.float64
+    dtype = np.float32
     print('Using NUMPY backend')
